+++ shared/utils/src/main/kotlin/ir/amirab/util/UrlUtils.kt
@@ -1,5 +1,8 @@
 package ir.amirab.util
 
+import ir.amirab.util.platform.Platform
+import java.awt.Desktop
+import java.net.URI
 import java.net.URL
 import java.net.URLDecoder
 
@@ -10,10 +13,10 @@ object UrlUtils {
         }.map { url ->
             val foundName = url.path
                 .split("/")
-                .filter { it.isNotBlank()}
+                .filter { it.isNotBlank() }
                 .lastOrNull()?.let {
                     kotlin.runCatching {
-                        URLDecoder.decode(it,Charsets.UTF_8)
+                        URLDecoder.decode(it, Charsets.UTF_8)
                     }.getOrNull()
                 }
             if (foundName != null) {
@@ -22,5 +25,14 @@ object UrlUtils {
             url.host.replace('.', '_')
         }
             .getOrNull()
+    }
+
+    fun openUrl(url: String) {
+        kotlin.runCatching {
+            val desktop = Desktop.getDesktop()
+            if (desktop.isSupported(Desktop.Action.BROWSE)) {
+                desktop.browse(URI(url))
+            }
+        }
     }
 }